set the thread instance to NULL after deleting
Fix memory errors when doing zlog_init/zlog_fini multiple times 
in the same process. This sets the thread instance data to NULL to
make sure that it gets recreated correctly.

This is useful in cases where multiple unit tests are 
performed in the same state if zlog is used in a python c module.