Added API to see if a given level is enabled
This commit adds an API to check if a given level is enabled before
build the message. Macros for default levels is also provided.

Sometimes the effort to build a complex/detailed log message can be very
costly, so it is useful in some applications to check if some level is
enabled in a category before take the decision to build the message.

Signed-off-by: Filipe Utzig <filipe.utzig@datacom.ind.br>