Implement UTC Time support
It is implemented using the ideas and remarks of @zaxbbun and @deemar in

It introduces two new format specifiers `%g` and `%G` (for GMT)
analougus to `%d` and `%D`. Unfurtunately the  `%D` format specifier is
already used as a shortform of `%d (%F %D)` and `%U` is use for the
function name.