conf: Return an error if zlog_init is called with a non-regular file
Currently, calling zlog_init with a non-regular file (e.g., a directory)
didn’t return an error. The logger would have been configured the same
as with an empty file.

The file is now checked to ensure it is regular, and an error is returned
if it is not.

Fixes #278.