CMake: append to CFLAGS instead of overwriting
There might be already some CFLAGS defined for CMake, which usually
shouldn't be overwritten (e.g. target specific optimizations or
--file-prefix arguments).

Instead of overwriting the CFLAGS in the CMakeLists.txt file, append
the flags to the existing list - in case there are some conflicting
flags present, the last ones will prevail.

Signed-off-by: Gyorgy Sarvari <skandigraun@gmail.com>